requirements.txt
boto3==1.16.35
botocore==1.19.35
Django==3.1.14
django-storages==1.10.1
djangorestframework==3.12.2
Pillow==9.3.0
python-decouple
qrcode
django-cors-headers
psycopg2==2.9.3

# opencv-python==4.5.1.48
pyyaml==5.4.1
# tensorflow-gpu==2.4.1 # install with conda
scipy==1.5.4
pandas==1.1.5
matplotlib==3.3.4
regex
requests
django_otp
